Move MessageSerializer implementation to serialization
