+++ serialization/src/androidMain/kotlin/com/boswelja/watchconnection/serialization/PrimitiveSerializers.kt
@@ -1,6 +1,5 @@
 package com.boswelja.watchconnection.serialization
 
-import com.boswelja.watchconnection.common.message.MessageSerializer
 import java.nio.ByteBuffer
 
 public actual class ShortSerializer actual constructor(
@@ -8,7 +7,7 @@ public actual class ShortSerializer actual constructor(
 ) : MessageSerializer<Short>(messagePaths) {
     override suspend fun serialize(data: Short): ByteArray =
         ByteBuffer.allocate(Short.SIZE_BYTES).putShort(data).array()
-    override suspend fun deserialize(bytes: ByteArray?): Short = ByteBuffer.wrap(bytes).short
+    override suspend fun deserialize(bytes: ByteArray?): Short = ByteBuffer.wrap(bytes!!).short
 }
 
 public actual class IntSerializer actual constructor(
@@ -16,7 +15,7 @@ public actual class IntSerializer actual constructor(
 ) : MessageSerializer<Int>(messagePaths) {
     override suspend fun serialize(data: Int): ByteArray =
         ByteBuffer.allocate(Int.SIZE_BYTES).putInt(data).array()
-    override suspend fun deserialize(bytes: ByteArray?): Int = ByteBuffer.wrap(bytes).int
+    override suspend fun deserialize(bytes: ByteArray?): Int = ByteBuffer.wrap(bytes!!).int
 }
 
 public actual class LongSerializer actual constructor(
@@ -24,7 +23,7 @@ public actual class LongSerializer actual constructor(
 ) : MessageSerializer<Long>(messagePaths) {
     override suspend fun serialize(data: Long): ByteArray =
         ByteBuffer.allocate(Long.SIZE_BYTES).putLong(data).array()
-    override suspend fun deserialize(bytes: ByteArray?): Long = ByteBuffer.wrap(bytes).long
+    override suspend fun deserialize(bytes: ByteArray?): Long = ByteBuffer.wrap(bytes!!).long
 }
 
 public actual class FloatSerializer actual constructor(
@@ -32,7 +31,7 @@ public actual class FloatSerializer actual constructor(
 ) : MessageSerializer<Float>(messagePaths) {
     override suspend fun serialize(data: Float): ByteArray =
         ByteBuffer.allocate(Float.SIZE_BYTES).putFloat(data).array()
-    override suspend fun deserialize(bytes: ByteArray?): Float = ByteBuffer.wrap(bytes).float
+    override suspend fun deserialize(bytes: ByteArray?): Float = ByteBuffer.wrap(bytes!!).float
 }
 
 public actual class DoubleSerializer actual constructor(
@@ -41,7 +40,7 @@ public actual class DoubleSerializer actual constructor(
     override suspend fun serialize(data: Double): ByteArray =
         ByteBuffer.allocate(Double.SIZE_BYTES).putDouble(data).array()
 
-    override suspend fun deserialize(bytes: ByteArray?): Double = ByteBuffer.wrap(bytes).double
+    override suspend fun deserialize(bytes: ByteArray?): Double = ByteBuffer.wrap(bytes!!).double
 }
 
 public actual class CharSerializer actual constructor(
@@ -49,7 +48,7 @@ public actual class CharSerializer actual constructor(
 ) : MessageSerializer<Char>(messagePaths) {
     override suspend fun serialize(data: Char): ByteArray =
         ByteBuffer.allocate(Char.SIZE_BYTES).putChar(data).array()
-    override suspend fun deserialize(bytes: ByteArray?): Char = ByteBuffer.wrap(bytes).char
+    override suspend fun deserialize(bytes: ByteArray?): Char = ByteBuffer.wrap(bytes!!).char
 }
 
 public actual class StringSerializer actual constructor(
